pyproject.toml
[tool.poetry]
name = "langchain-prac"
version = "0.1.0"
description = ""
authors = ["Harshit <harshitjain5june@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.1.16"
langchain-openai = "^0.1.4"
python-dotenv = "^1.0.1"
beautifulsoup4 = "^4.12.3"
faiss-cpu = "^1.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
